Full simulation of the project is now working

+++ src/Data Structures/CancelledRequest.h
@@ -3,7 +3,7 @@
 
 struct CancelledRequest
 {
-    int cTimestep;
+    double cTimestep;
     int cPID;
     int cHID;
 };
